Move monkey patch into separate file
